Start reorganizing bin/parse-gameday-xml into modules
I goal is to be able to type something like:

  from mlbam import gameday
  gameday.fetch(...)
  gameday.parse('some/dir')
  ...
  gameday.pitchfx....

The fetch can be moved without too much trouble, but the parsing will
take a little while. The method now named _parse() is huge and needs to
be split into several smaller methods. Additionally, it won't interact
with the database at all (I think..). This could make importing mass
amounts of data tedious, so I could make it directly import the data.