Update all park hp_x and hp_y and set the scale for 12-16
All of the parks' hp_x values are the same now (woohoo!) with only one
park having a different hp_y. I will need to check that one when I get
to it.
Also, updated the scale factors for parks 12-16. This had already been
done and wasn't worth the effort temporarily reverting that.