Add function to create dtype for Retrosheet game info
I think this is the right way to do it? Need to add the extended fields
still. This would also be a good spot to parse the event data. It might
be useful to flip this around into a generator and have the loop consume
each game as its parsed.