Adjust how the bip are associated with the plate appearance
Multiple "bip" can be associated with one plate appearance; e.g. a hit
and error on the play. This means each successful match needs to start
again with the same plate appearance and the next bip.

Further work is needed to support one plate appearance having multiple
bip associated with it. The database will be confused right now seeing
a list instead of an integer.