Start new set of functions to parse cwgame and cwevent simultaneously
Rather than parsing the event files once with cwgame and then again with
cwevent, parse each event file simultaneously and merge the event data
with the game data and return that as a generator. Each iteration would
then return a new game that can be saved in the HDF5 file.